+++ src/modules/attestation/services/RevocationService.ts
@@ -16,10 +16,11 @@
  */
 
 import { account, ChainId, DeleteAccountPropertyParams, DeleteAccountPropertyResponse, IRequest } from "@somedotone/ardor-ts";
-import { EntityType, ErrorCode, objectAny } from "../../..";
-import { AttestationResponse, IAttestationService } from "../../../types";
+import { EntityType, ErrorCode } from "../../..";
+import { AttestationResponse } from "../../../types";
 import DataFields from "../../lib/DataFields";
 import Helper from "../../lib/Helper";
+import { IAttestationService, objectAny } from "../../internal-types";
 import ServiceHelper from "./utils/ServiceHelper";
